Persona_Genie_Generic/persona_genie/requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
matplotlib==3.8.2
wordcloud==1.9.2
pillow==10.2.0
faker==22.0.0
crewai==0.28.0
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0 